Move javascript include body tag below body tag
If the generated javascript needs to access the DOM for injection or
checking then it will return a Not Defined error because the DOM has not
rendered yet. Now if we have DOM manipulation it will always be after
the DOM has rendered properly for manipulation. Test this by
manipulating the DOM inside site.js