Refactor entry and migrate to scene file
The previous entry js file was convoluted and difficult to add extra
javascript to, now the webvr/threejs is in its own scene file for
portability and readability. The reason we had the previous entry was
for prototyping, but now our goal is for 3D VR chess which makes the new
entry and seperation apt.