Add Chess class
Chess class is the center point of a chess game, it is initalized then
can be invoked to play using the play() method which listens to a
websocket for state changes. Chess has access to the scene to allow
rendering on the fly. Has two players for keeping track of metadata.