Add pre-push linting
Before, linting errors would sneak into master due to not having builds
running or not using linting consistently during our development
process. Now we have a pre-push hook that runs linting on our
front-end code and if any errors happen pushing is aborted and  we are prompted to fix them or
force push them anyways. This allows us to ensure when we push we are
alerting to check linting.