Add meta tag for hiding address bar
Before on iOS if we demo the application through safari the address bar
will always be present. Now if we add the site to our home screen we can
hide the address bar for a better experience. Once we have a dev server
up with a consistent URL this will be ideal.