Add fix-lint npm command
Before we would have to fix errors by hand or run eslint from the
command line. Now we have it wrapped in a simple package script.
fix-lint will run eslint with the --fix flag which tries to fix any
linting errors on its own. This is helpful when a lot of linting errors
are due to inconsistent whitespace, for example. This closes #48.