Fix middleman build minify failure
Currently the build command of middleman fails, which does not allow us
to build minified css/js. The reason for the failure is the main
UglifyJS does not have support for
[es6](https://github.com/webpack-contrib/uglifyjs-webpack-plugin). The
fix is to use uglify-js(es6 support) and not the default webpack uglifyjs.
Since removing the webpack default compression(-p flag), we  add the
minify-css flag to middlemans config file to minify the css.