Use a shader for the board
Add BoardShader.js to use it as a material for the Board.
Squares are no longer used in the board class.
The positioning methods for the pieces has been rewritten due to the
changes in Board.
Changes have been made to Chess.js to no longer handle a promise for
Board.js.