Add material to chess pieces
Before only the knight had materials applied to it, therefore the others
would not have shine or light truly alter the aesthetic of the piece.
The solution: add and apply the material from the knight to the
remaining pieces. Now all blender files of the pieces have a material
for loading.