Add stats.js for FPS monitoring
Monitoring the FPS is important to see if any performance hits are taken
if we modify how we render the pieces, board or any of the logic. It is
implemented to show in mobile as well. Stats.js is used due to its
integration with three.js. It wraps around the animation frame loop.