Remove middleman
Our chessvr is a strictly javascript application, without the need for
erb or ruby stuff we migrated to webpack only for our build process.
Alongside removing middleman some cleanup was done on the webpack side
of things and we decreased build time from 24 seconds on average to 300ms on average.